Add small api for IP validation of members
Kamailio can use an API call to validate a source IP as a member of a
given dispatcher set.  This can be an easier, more direct way of
establishing source IP trust than separately maintaining that table.

If enabled (via the -api flag), the service will listen for HTTP
requests to `/check/<setID>/<ip address>` and return a 200 OK if and
only if that IP address is a member of that dispatcher set.

This also converts to go modules.